exercise 0: second method
The objects are send as a reference. This makes that the value of the parameter `other` in the `secondMethod` is override with the new reference of the new instance created inside the method (`new SomeType()`). So, all the changes done to the new instance do not affect to the original one.