exercise 0: third method
As we mention in the previous step the objects are send as a reference through the functions. So, you have a pointer to a piece of memory where is stored the original object outside the function. Knowing that, all the changes that you made directly to the reference will affect outside the scope of the method.