exercise 0: fourth method
In this last exercise you will know the meaning of the `this` keyword. `This` means the context of the object over the method has been call. So, in this sample `this` is equal to `instance` because you call the fourth method in this way `instance.fourthMethod`. Knowing that, `instance` it has been created at the beginning of the test and it is never modified and that the `int` type has the 0 as the default value.
For that reason, when inside the method the array of the instance send through parameter is overridden by the instance of the `this` context is set to 0 again.