exercise 2.2: inline the old implementation.
Put the caret in the name of the old implementation `isAuthenticated(int id)` and do an inline for this method. Now, all the old implementations will be replaced with the code that had inside of it. In this case, the new implementation that we want to use.

Doing that, we will not need to go one by one replacing all the usages.